+++ core/src/ten_manager/designer_frontend/src/components/ui/Table.tsx
