+++ core/src/ten_manager/designer_frontend/src/components/ui/Tabs.tsx
